requirements.txt
cycler==0.10.0
Django==2.0.6
django-cors-headers==2.3.0
django-extensions==2.0.7
django-filter==2.0.0
djangorestframework==3.8.2
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.0
numpy==1.15.1
pandas==0.23.4
Pint==0.8.1
psycopg2==2.7.5
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
xlrd==1.1.0
XlsxWriter==1.1.1
xmltodict==0.11.0
